Move tests to chia/tests
